+++ src/main/java/io/github/mfvanek/pg/checks/cluster/ForeignKeysNotCoveredWithIndexCheckOnCluster.java
@@ -12,7 +12,7 @@ package io.github.mfvanek.pg.checks.cluster;
 
 import io.github.mfvanek.pg.checks.host.ForeignKeysNotCoveredWithIndexCheckOnHost;
 import io.github.mfvanek.pg.connection.HighAvailabilityPgConnection;
-import io.github.mfvanek.pg.model.index.ForeignKey;
+import io.github.mfvanek.pg.model.constraint.ForeignKey;
 
 import javax.annotation.Nonnull;
