+++ src/test/java/io/github/mfvanek/pg/model/column/SerialTypeTest.java
@@ -8,9 +8,8 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.model.table;
+package io.github.mfvanek.pg.model.column;
 
-import io.github.mfvanek.pg.model.column.SerialType;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
