+++ src/test/java/io/github/mfvanek/pg/model/constraint/ForeignKeyTest.java
@@ -8,7 +8,7 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.model.index;
+package io.github.mfvanek.pg.model.constraint;
 
 import io.github.mfvanek.pg.model.column.Column;
 import nl.jqno.equalsverifier.EqualsVerifier;
